Image-/requirements.txt
pyTelegramBotAPI
Pillow
flask
gunicorn
